requirements.txt
# Core
pyyaml>=6.0

# Optional — only needed by workflows that use LLMs directly
# openai>=1.0.0
